Fully implemented a simple storage system
Using the "testStorage.html" file, I now have a working load function that will pull every entry from the "apps.json" storage file.

Also added some classes to "style.css" for the test page and added a link to the test page in "index.html" to easily navigate to it.