Major UI Update 1
Changed the background to a gradient created by code rather than a picture. All pages are fully connected with animated button/links. Added descriptions to the Home page and About page. Added a search button, but as of right now doesn't work. Added different font (golos and pacifico) for sleeker look.
Need to fix the width of the header and footer, as it gets cut off for some reason.